requirements_dev.txt
invoke
twine
wheel
zest.releaser
black
isort
pytest
pytest-django
pytest-djangoapp
